+++ CHANGELOG.md
@@ -19,6 +19,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - New role flag: `permanent = true`.
   Permanent roles are always enabled. Also they are hidden implicitly.
+  
+- New functions in cluster test_helpers - `Cluster:upload_config(config)` and `Cluster:download_config()`  
 
 ### Fixed